readyup|refactor: Import JSON payload types from the modules that define them
Importing a single JSON payload type no longer pulls in every schema module.
