+++ packages/readyup/.readyup/kits/__tests__/bundledKits.tool.test.ts
@@ -11,8 +11,9 @@ import { compileConfig } from '../../../src/compile/compileConfig.ts';
 import { validateCompiledOutput } from '../../../src/compile/validateCompiledOutput.ts';
 import { listCommand } from '../../../src/list/listCommand.ts';
 import type { RdyManifestKit } from '../../../src/manifest/manifestSchema.ts';
-import type { JsonKitResultEntry, JsonReport } from '../../../src/schemas/index.ts';
-import { ListOutputSchema, ReportSchema } from '../../../src/schemas/index.ts';
+import { ListOutputSchema } from '../../../src/schemas/listOutputSchema.ts';
+import type { JsonKitResultEntry, JsonReport } from '../../../src/schemas/reportSchema.ts';
+import { ReportSchema } from '../../../src/schemas/reportSchema.ts';
 import { hashFile } from '../../../src/verify/targetHash.ts';
 
 const PACKAGE_ROOT = path.resolve(import.meta.dirname, '..', '..', '..');
