+++ packages/readyup/src/compile/compileCommand.ts
@@ -11,8 +11,7 @@ import { DEFAULT_MANIFEST_PATH } from '../manifest/manifestPath.ts';
 import type { RdyManifestKit } from '../manifest/manifestSchema.ts';
 import { ManifestNotFoundError, readManifest } from '../manifest/readManifest.ts';
 import { writeManifest } from '../manifest/writeManifest.ts';
-import { SCHEMA_VERSION } from '../schemas/compileOutputSchema.ts';
-import type { JsonCompileKitEntry, JsonCompileOutput } from '../schemas/index.ts';
+import { type JsonCompileKitEntry, type JsonCompileOutput, SCHEMA_VERSION } from '../schemas/compileOutputSchema.ts';
 import { extractHint, extractMessage } from '../utils/error-handling.ts';
 import { translateParseArgsError } from '../utils/parse-args-error.ts';
 import { pluralizeWithCount } from '../utils/pluralize.ts';
