+++ packages/readyup/src/list/listCommand.ts
@@ -23,8 +23,7 @@ import { loadRemoteManifest } from '../remote/loadRemoteManifest.ts';
 import { resolveRemoteAuthHeaders, resolveRemoteProvider } from '../remote/remote-provider.ts';
 import { toRemoteRdyError } from '../remote/toRemoteRdyError.ts';
 import { resolvePackageRoot } from '../resolvePackageRoot.ts';
-import type { JsonListKitEntry, JsonListOutput } from '../schemas/index.ts';
-import { SCHEMA_VERSION } from '../schemas/listOutputSchema.ts';
+import { type JsonListKitEntry, type JsonListOutput, SCHEMA_VERSION } from '../schemas/listOutputSchema.ts';
 import { extractHint, extractMessage } from '../utils/error-handling.ts';
 import { translateParseArgsError } from '../utils/parse-args-error.ts';
 import { writeHuman } from '../writeHuman.ts';
