+++ packages/readyup/src/verify/verifyCommand.ts
@@ -11,8 +11,7 @@ import type { TokenName } from '../layout/formatter.ts';
 import { DEFAULT_MANIFEST_PATH } from '../manifest/manifestPath.ts';
 import type { RdyManifestKit } from '../manifest/manifestSchema.ts';
 import { readManifest } from '../manifest/readManifest.ts';
-import type { JsonVerifyKitEntry, JsonVerifyOutput } from '../schemas/index.ts';
-import { SCHEMA_VERSION } from '../schemas/verifyOutputSchema.ts';
+import { type JsonVerifyKitEntry, type JsonVerifyOutput, SCHEMA_VERSION } from '../schemas/verifyOutputSchema.ts';
 import { extractMessage } from '../utils/error-handling.ts';
 import { translateParseArgsError } from '../utils/parse-args-error.ts';
 import { VERSION } from '../version.ts';
